state and state_resource helpers
