+++ lib/rspec/terraspace/ts.rb
@@ -3,6 +3,7 @@ require "json"
 module RSpec::Terraspace
   class Ts
     extend Memoist
+    include Concern
 
     CLI = ::Terraspace::CLI
 
@@ -16,7 +17,6 @@ module RSpec::Terraspace
       run("up #{args} -y")
       mod = args.split(' ').first
       @mod = ::Terraspace::Mod.new(mod)
-      save_output
     end
 
     def down(args)
@@ -27,24 +27,6 @@ module RSpec::Terraspace
       puts "=> TS_ENV=#{Terraspace.env} terraspace #{command}".color(:green)
       args = command.split(' ')
       CLI.start(args)
-    end
-
-    # Note: a terraspace.down will remove the output.json since it does a clean
-    def save_output
-      FileUtils.mkdir_p(File.dirname(out_path))
-      run("output #{@mod.name} --format json --out #{out_path}")
-    end
-
-    def output(mod, name)
-      outputs.dig(name, "value")
-    end
-
-    def outputs
-      JSON.load(IO.read(out_path))
-    end
-
-    def out_path
-      "#{Terraspace.tmp_root}/rspec/output.json"
     end
   end
 end
